Completed add of NeoPixel destructor
The entire purpose for doing this is so that I can have a main loop that re-constructs a set of led instructions based upon input from a file. If the details in the file change, those should be loaded. This means re-reading the file, and reconstructing a new Adafruit_NeoPixel() object each time... only problem was that until now, the delete never really deleted anything.