Explictly initialize member variables brightness and endTime to default values (0)
In certain circumstances, namely as member variable in a compound
class, or as a base class, of a object instantiated on the heap, the
default value can not be guaranteed to be zero. In both cases, a
brightness or endTime of indeterminate value could cause unexpected
and/or undefined behavior.