Color Wheel Fix
The supposed R-G-B spectrum of the Wheel method was all wrong. It started with a green, then went to yellow, then red, then purple, then blue. I've completely fixed this, and tested it on my neopixel strands. This new Wheel method starts at pure red, moves to orange, yellow, green, and the full rainbow, as it is in nature.