Support of BBC micro:bit
Microsoft PXT/neopixel (JavaScript?) sendBufferASM function that runs the neopixels on the make:code website. (https://github.com/Microsoft/pxt-neopixel/blob/master/sendbuffer.asm)
was ported to the arduino IDE asm syntax, inserted just before

#else // Other ARM architecture -- Presumed Arduino Due
starting with
#elif defined (VARIANT_BBC_MICROBIT)

The original sendBufferASM is commented out (for verification). Very ugly now. I'm happy to clean it up but I thought this might help someone with more advanced skills verify my tinkering.

Differences between the new code and sendBuffer

the assignment of the bitmask, following
https://github.com/sandeepmistry/arduino-nRF5/blob/dc53980c8bac27898fca90d8ecb268e11111edc1/variants/BBCmicrobit/variant.cpp

the assignment of the *reg (similar to the other ARM architecture), following
https://github.com/sandeepmistry/arduino-nRF5/blob/dc53980c8bac27898fca90d8ecb268e11111edc1/cores/nRF5/SDK/components/device/nrf51.h
and
http://www.iot-programmer.com/index.php/books/27-micro-bit-iot-in-c/chapters-micro-bit-iot-in-c/47-micro-bit-iot-in-c-fast-memory-mapped-gpio?showall=1

lsrs, adds and subs in sendBufferASM were not supported; changed to lsr, add and sub (with no detrimental consequenses)

The ported code was tested with example "strandtest" and a 150 rgb led string on pins 1 and 2 of the micro:bit.