Class formatting; no functional change
Apparently my "group all the variables of the same type with a single declaration" style of formatting upsets some and is considered distasteful...so, okay, they're all broken out to their own items now. This is just a code style thing and makes no functional difference in compiled code, so I'm not bumping the version number.