Change pin type in constructors & setPin() to int16_t
Internally the pin # is a signed type anyway (negative = no output pin assigned), this just extends that to the constructor. Only some parts of the code handle this correctly and it would be wise to make a pass through everything to check and correct any oversights. Type is intentionally 16-bit because some boards have huge pin numbers.