ESP32: Using RGB_BUILTIN define as pin number in the constructor fails and resets S2 and S3
I was experimenting with a few different ESP32 boards (C3, S2, S3) and was playing with the on board neopixel and got tired of changing the pin number (C3->8, S2->18, S3->48), and noticed they had a define RGB_BUILTIN and hoped that you could use it in the constructor.

The neopixel failed to work, and on the S2 and S3 it guru-meditated.  Found out they set these values as the pin number plus the count of pins...

So in setPin I checked for this and offset back to the actual pin number.