cope with I$ miss on teensy
if the instruction cache is completely flushed *even before* the counter
is read, under specific situations (presumably when relevant parts of
show() are on different cache lines), the first bit would not be sent
(but discarded), often leading to a change in colour.

prevent compiler optimisation across counter reads,
order the delays so that they are easier to reason about.